requirements.txt
Pillow
opencv-python
numpy
python3_xlib
pyautogui
pywin32
ipython
pytesseract
coverage
keyboard
